backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
ecdsa==0.19.1
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
bandit==1.8.3
# semgrep � not supported on Windows. Use WSL: pip install semgrep
GitPython==3.1.44
requests==2.32.3
beautifulsoup4==4.13.3
motor==3.7.0
pymongo==4.10.1
reportlab==4.3.1
joblib==1.4.2
scikit-learn==1.7.2
numpy==2.2.3
pandas==2.2.3
slowapi==0.1.9 
stripe==15.0.1
apscheduler==3.10.4